fix: 사물함 대여 시 모달이 열리지 않는 현상 수정
