+++ frontend/src/mocks/data/user.json
@@ -4,7 +4,7 @@
     "name": "이지우",
     "id": "C211000",
     "phoneNumber": "010-1234-5678",
-    "role": "GENERAL",
+    "role": "PRESIDENT",
     "major": "컴퓨터공학과",
     "accessToken": "auth"
   }
